+++ mocks/mock_API.go
@@ -8,6 +8,7 @@ import (
 	"context"
 
 	mock "github.com/stretchr/testify/mock"
+	"github.com/winning-number/fdj-sdk/v2"
 	"github.com/winning-number/fdj-sdk/v2/source"
 )
 
@@ -100,6 +101,74 @@ func (_c *MockAPI_DownloadHistory_Call) Return(source1 source.Source, err error)
 }
 
 func (_c *MockAPI_DownloadHistory_Call) RunAndReturn(run func(ctx context.Context, datasetID string) (source.Source, error)) *MockAPI_DownloadHistory_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
+// Draws provides a mock function for the type MockAPI
+func (_mock *MockAPI) Draws(ctx context.Context, filter *fdj.DrawFilter) (*fdj.DrawData, error) {
+	ret := _mock.Called(ctx, filter)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Draws")
+	}
+
+	var r0 *fdj.DrawData
+	var r1 error
+	if returnFunc, ok := ret.Get(0).(func(context.Context, *fdj.DrawFilter) (*fdj.DrawData, error)); ok {
+		return returnFunc(ctx, filter)
+	}
+	if returnFunc, ok := ret.Get(0).(func(context.Context, *fdj.DrawFilter) *fdj.DrawData); ok {
+		r0 = returnFunc(ctx, filter)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*fdj.DrawData)
+		}
+	}
+	if returnFunc, ok := ret.Get(1).(func(context.Context, *fdj.DrawFilter) error); ok {
+		r1 = returnFunc(ctx, filter)
+	} else {
+		r1 = ret.Error(1)
+	}
+	return r0, r1
+}
+
+// MockAPI_Draws_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'Draws'
+type MockAPI_Draws_Call struct {
+	*mock.Call
+}
+
+// Draws is a helper method to define mock.On call
+//   - ctx context.Context
+//   - filter *fdj.DrawFilter
+func (_e *MockAPI_Expecter) Draws(ctx interface{}, filter interface{}) *MockAPI_Draws_Call {
+	return &MockAPI_Draws_Call{Call: _e.mock.On("Draws", ctx, filter)}
+}
+
+func (_c *MockAPI_Draws_Call) Run(run func(ctx context.Context, filter *fdj.DrawFilter)) *MockAPI_Draws_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		var arg0 context.Context
+		if args[0] != nil {
+			arg0 = args[0].(context.Context)
+		}
+		var arg1 *fdj.DrawFilter
+		if args[1] != nil {
+			arg1 = args[1].(*fdj.DrawFilter)
+		}
+		run(
+			arg0,
+			arg1,
+		)
+	})
+	return _c
+}
+
+func (_c *MockAPI_Draws_Call) Return(drawData *fdj.DrawData, err error) *MockAPI_Draws_Call {
+	_c.Call.Return(drawData, err)
+	return _c
+}
+
+func (_c *MockAPI_Draws_Call) RunAndReturn(run func(ctx context.Context, filter *fdj.DrawFilter) (*fdj.DrawData, error)) *MockAPI_Draws_Call {
 	_c.Call.Return(run)
 	return _c
 }